move pkg/scheduler/util/testutil.go to pkg/scheduler/testing
